+++ datamodel/odata-v4/odata-v4-api-sample/src/main/java/com/sap/cloud/sdk/datamodel/odatav4/sample/namespaces/sdkgrocerystore/Receipt.java
@@ -273,61 +273,61 @@ public class Receipt extends VdmEntity<Receipt> implements VdmEntitySet
         // simple properties
         {
             if( cloudSdkValues.containsKey("Id") ) {
-                final Object value = cloudSdkValues.remove("Id");
-                if( (value == null) || (!value.equals(getId())) ) {
-                    setId(((Integer) value));
+                final Object cloudSdkValue = cloudSdkValues.remove("Id");
+                if( (cloudSdkValue == null) || (!cloudSdkValue.equals(getId())) ) {
+                    setId(((Integer) cloudSdkValue));
                 }
             }
             if( cloudSdkValues.containsKey("CustomerId") ) {
-                final Object value = cloudSdkValues.remove("CustomerId");
-                if( (value == null) || (!value.equals(getCustomerId())) ) {
-                    setCustomerId(((Integer) value));
+                final Object cloudSdkValue = cloudSdkValues.remove("CustomerId");
+                if( (cloudSdkValue == null) || (!cloudSdkValue.equals(getCustomerId())) ) {
+                    setCustomerId(((Integer) cloudSdkValue));
                 }
             }
             if( cloudSdkValues.containsKey("TotalAmount") ) {
-                final Object value = cloudSdkValues.remove("TotalAmount");
-                if( (value == null) || (!value.equals(getTotalAmount())) ) {
-                    setTotalAmount(((BigDecimal) value));
+                final Object cloudSdkValue = cloudSdkValues.remove("TotalAmount");
+                if( (cloudSdkValue == null) || (!cloudSdkValue.equals(getTotalAmount())) ) {
+                    setTotalAmount(((BigDecimal) cloudSdkValue));
                 }
             }
         }
         // structured properties
         {
             if( cloudSdkValues.containsKey("ProductCounts") ) {
-                final Object value = cloudSdkValues.remove("ProductCounts");
-                if( value instanceof Iterable ) {
+                final Object cloudSdkValue = cloudSdkValues.remove("ProductCounts");
+                if( cloudSdkValue instanceof Iterable ) {
                     final LinkedList<ProductCount> productCounts = new LinkedList<ProductCount>();
-                    for( Object cloudSdkProperties : ((Iterable<?>) value) ) {
+                    for( Object cloudSdkProperties : ((Iterable<?>) cloudSdkValue) ) {
                         if( cloudSdkProperties instanceof Map ) {
-                            final ProductCount item = new ProductCount();
+                            final ProductCount cloudSdkItem = new ProductCount();
                             @SuppressWarnings( "unchecked" )
-                            final Map<String, Object> inputMap = ((Map<String, Object>) value);
-                            item.fromMap(inputMap);
-                            productCounts.add(item);
+                            final Map<String, Object> inputMap = ((Map<String, Object>) cloudSdkValue);
+                            cloudSdkItem.fromMap(inputMap);
+                            productCounts.add(cloudSdkItem);
                         }
                     }
                     setProductCounts(productCounts);
                 }
-                if( (value == null) && (getProductCounts() != null) ) {
+                if( (cloudSdkValue == null) && (getProductCounts() != null) ) {
                     setProductCounts(null);
                 }
             }
             if( cloudSdkValues.containsKey("properties") ) {
-                final Object value = cloudSdkValues.remove("properties");
-                if( value instanceof Iterable ) {
+                final Object cloudSdkValue = cloudSdkValues.remove("properties");
+                if( cloudSdkValue instanceof Iterable ) {
                     final LinkedList<ProductCount> properties = new LinkedList<ProductCount>();
-                    for( Object cloudSdkProperties : ((Iterable<?>) value) ) {
+                    for( Object cloudSdkProperties : ((Iterable<?>) cloudSdkValue) ) {
                         if( cloudSdkProperties instanceof Map ) {
-                            final ProductCount item = new ProductCount();
+                            final ProductCount cloudSdkItem = new ProductCount();
                             @SuppressWarnings( "unchecked" )
-                            final Map<String, Object> inputMap = ((Map<String, Object>) value);
-                            item.fromMap(inputMap);
-                            properties.add(item);
+                            final Map<String, Object> inputMap = ((Map<String, Object>) cloudSdkValue);
+                            cloudSdkItem.fromMap(inputMap);
+                            properties.add(cloudSdkItem);
                         }
                     }
                     setProperties(properties);
                 }
-                if( (value == null) && (getProperties() != null) ) {
+                if( (cloudSdkValue == null) && (getProperties() != null) ) {
                     setProperties(null);
                 }
             }